todo/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.8
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
excel2json-3==0.1.6
Faker==12.0.1
idna==3.3
importlib-metadata==4.11.2
Markdown==3.3.6
mixer==7.2.2
notes==0.3.0
openpyxl==3.0.9
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
xlrd==2.0.1
zipp==3.7.0
